+++ src/app-toolkit/decorators/index.ts
@@ -2,7 +2,6 @@ import { AppDefinition, AppModule } from '~app/app.decorator';
 import { ContractType } from '~position/contract.interface';
 import { PositionBalanceFetcher } from '~position/position-balance-fetcher.decorator';
 import { PositionFetcher } from '~position/position-fetcher.decorator';
-import { TvlFetcher } from '~stats/tvl/tvl-fetcher.decorator';
 
 import { BalanceFetcher } from './balance-fetcher.decorator';
 import { BalancePresenter } from './balance-presenter.decorator';
@@ -16,5 +15,4 @@ export const Register = {
   TokenPositionFetcher: PositionFetcher(ContractType.APP_TOKEN),
   ContractPositionBalanceFetcher: PositionBalanceFetcher(ContractType.POSITION),
   TokenPositionBalanceFetcher: PositionBalanceFetcher(ContractType.APP_TOKEN),
-  TvlFetcher,
 };