+++ src/apps/badger/ethereum/badger.vault.token-fetcher.ts
@@ -31,12 +31,12 @@ export type BadgerApiTokensResponseEntry = {
 
 export type BadgerApiTokensResponse = Record<string, BadgerApiTokensResponseEntry>;
 
-@Register.TokenPositionFetcher({ appId, groupId, network, options: { includeInTvl: true } })
+@Register.TokenPositionFetcher({ appId, groupId, network })
 export class EthereumBadgerVaultTokenFetcher implements PositionFetcher<AppTokenPosition> {
   constructor(@Inject(BadgerVaultTokenHelper) private readonly badgerVaultTokenHelper: BadgerVaultTokenHelper) {}
 
   @CacheOnInterval({
-    key: `apps-v3:${network}:${appId}:${groupId}:definitions`,
+    key: `studio:${appId}:${groupId}:${network}:definitions`,
     timeout: 15 * 60 * 1000,
   })
   async getVaultDefinitions() {
@@ -55,9 +55,9 @@ export class EthereumBadgerVaultTokenFetcher implements PositionFetcher<AppToken
       definitions: vaultDefinitions,
       dependencies: [
         { appId: CURVE_DEFINITION.id, groupIds: [CURVE_DEFINITION.groups.pool.id], network },
+        { appId: UNISWAP_V2_DEFINITION.id, groupIds: [UNISWAP_V2_DEFINITION.groups.pool.id], network },
         { appId: 'harvest', groupIds: ['vault'], network },
         { appId: 'sushiswap', groupIds: ['pool'], network },
-        { appId: UNISWAP_V2_DEFINITION.id, groupIds: [UNISWAP_V2_DEFINITION.groups.pool.id], network },
       ],
     });
   }