+++ src/apps/good-ghosting/celo/good-ghosting.game.contract-position-fetcher.ts
@@ -1,7 +1,5 @@
 import { Inject } from '@nestjs/common';
 
-import { IAppToolkit, APP_TOOLKIT } from '~app-toolkit/app-toolkit.interface';
-
 import { Register } from '~app-toolkit/decorators';
 import { PositionFetcher } from '~position/position-fetcher.interface';
 import { ContractPosition } from '~position/position.interface';
@@ -9,8 +7,6 @@ import { Network } from '~types/network.interface';
 
 import { GOOD_GHOSTING_DEFINITION } from '../good-ghosting.definition';
 import { NetworkId } from '../helpers/constants';
-import { GoodGhostingContractFactory } from '../contracts';
-
 import { GoodGhostingGameContractPositionFetcherHelper } from '../helpers/good-ghosting.game.contract-position-fetcher-helper';
 
 const appId = GOOD_GHOSTING_DEFINITION.id;
@@ -21,8 +17,6 @@ const networkId = NetworkId.CeloMainnet;
 @Register.ContractPositionFetcher({ appId, groupId, network })
 export class CeloGoodGhostingGameContractPositionFetcher implements PositionFetcher<ContractPosition> {
   constructor(
-    @Inject(GoodGhostingContractFactory) private readonly goodGhostingContractFactory: GoodGhostingContractFactory,
-
     @Inject(GoodGhostingGameContractPositionFetcherHelper)
     private readonly helper: GoodGhostingGameContractPositionFetcherHelper,
   ) {}